+++ src/features/schedules/hooks/useVehicleSchedulesData.tsx
@@ -1,14 +1,14 @@
 import { useState, useEffect, useCallback, useRef } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { Driver, DriverQueryParams, PaginatedDriversResponse } from '../types';
-import { driverService } from '../services/apiService';
+import { VehicleSchedule, VehicleScheduleQueryParams, PaginatedVehicleSchedulesResponse, VehicleScheduleSummary } from '../../../types';
+import { vehicleScheduleService } from '../../../services/apiService';
 
-type SortColumn = 'name' | 'email' | 'idNumber' | 'createdAt';
+type SortColumn = 'vehicleName' | 'driverName' | 'startDate' | 'endDate' | 'duration' | 'status';
 type SortDirection = 'asc' | 'desc';
 
-interface UseDriversDataReturn {
+interface UseVehicleSchedulesDataReturn {
   // Data
-  drivers: Driver[];
+  vehicleSchedules: VehicleSchedule[];
   totalCount: number;
   totalPages: number;
   currentPage: number;
@@ -17,7 +17,7 @@ interface UseDriversDataReturn {
   
   // State
   searchTerm: string;
-  emailSearchTerm: string;
+  statusFilters: string[];
   sortBy: SortColumn | null;
   sortOrder: SortDirection;
   itemsPerPage: number;
@@ -26,30 +26,36 @@ interface UseDriversDataReturn {
   
   // Actions
   setSearchTerm: (term: string) => void;
-  setEmailSearchTerm: (term: string) => void;
+  toggleStatusFilter: (status: string) => void;
+  clearStatusFilters: () => void;
   setSorting: (column: SortColumn) => void;
   setCurrentPage: (page: number) => void;
   setItemsPerPage: (limit: number) => void;
   clearAllFilters: () => void;
   refreshData: () => Promise<void>;
+
+  // Summary data
+  vehicleScheduleSummary: VehicleScheduleSummary | null;
 }
 
-export function useDriversData(): UseDriversDataReturn {
+export function useVehicleSchedulesData(): UseVehicleSchedulesDataReturn {
   const [searchParams, setSearchParams] = useSearchParams();
   
   // Initialize state from URL parameters
   const getInitialState = () => {
     const search = searchParams.get('search') || '';
-    const emailSearch = searchParams.get('emailSearch') || '';
+    const status = searchParams.getAll('status');
+    // If no status filters are specified in URL, use backend's default filters
+    const defaultStatusFilters = status.length > 0 ? status : ['active', 'scheduled'];
     const sortBy = searchParams.get('sortBy') as SortColumn | null;
-    const sortOrder = (searchParams.get('sortOrder') as SortDirection) || 'asc';
+    const sortOrder = (searchParams.get('sortOrder') as SortDirection) || 'desc'; // Default to newest first
     const page = parseInt(searchParams.get('page') || '1', 10);
-    const limit = parseInt(searchParams.get('limit') || '10', 10);
+    const limit = parseInt(searchParams.get('limit') || '6', 10);
     
     return {
       search,
-      emailSearch,
-      sortBy,
+      status: defaultStatusFilters,
+      sortBy: sortBy || 'startDate', // Default sort by start date
       sortOrder,
       page,
       limit,
@@ -60,28 +66,27 @@ export function useDriversData(): UseDriversDataReturn {
   
   // State variables
   const [searchTerm, setSearchTermState] = useState(initialState.search);
-  const [emailSearchTerm, setEmailSearchTermState] = useState(initialState.emailSearch);
+  const [statusFilters, setStatusFilters] = useState<string[]>(initialState.status);
   const [sortBy, setSortByState] = useState<SortColumn | null>(initialState.sortBy);
   const [sortOrder, setSortOrderState] = useState<SortDirection>(initialState.sortOrder);
   const [currentPage, setCurrentPageState] = useState(initialState.page);
   const [itemsPerPage, setItemsPerPageState] = useState(initialState.limit);
   
   // Data state
-  const [drivers, setDrivers] = useState<Driver[]>([]);
+  const [vehicleSchedules, setVehicleSchedules] = useState<VehicleSchedule[]>([]);
   const [totalCount, setTotalCount] = useState(0);
   const [totalPages, setTotalPages] = useState(0);
   const [hasNextPage, setHasNextPage] = useState(false);
   const [hasPreviousPage, setHasPreviousPage] = useState(false);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
+  const [vehicleScheduleSummary, setVehicleScheduleSummary] = useState<VehicleScheduleSummary | null>(null);
 
   // Debouncing for search
   const searchTimeoutRef = useRef<NodeJS.Timeout>();
-  const emailSearchTimeoutRef = useRef<NodeJS.Timeout>();
   const [debouncedSearchTerm, setDebouncedSearchTerm] = useState(searchTerm);
-  const [debouncedEmailSearchTerm, setDebouncedEmailSearchTerm] = useState(emailSearchTerm);
 
-  // Debounce search terms
+  // Debounce search term
   useEffect(() => {
     if (searchTimeoutRef.current) {
       clearTimeout(searchTimeoutRef.current);
@@ -98,35 +103,19 @@ export function useDriversData(): UseDriversDataReturn {
     };
   }, [searchTerm]);
 
-  useEffect(() => {
-    if (emailSearchTimeoutRef.current) {
-      clearTimeout(emailSearchTimeoutRef.current);
-    }
-    
-    emailSearchTimeoutRef.current = setTimeout(() => {
-      setDebouncedEmailSearchTerm(emailSearchTerm);
-    }, 300); // 300ms debounce
-
-    return () => {
-      if (emailSearchTimeoutRef.current) {
-        clearTimeout(emailSearchTimeoutRef.current);
-      }
-    };
-  }, [emailSearchTerm]);
-
   // Update URL parameters when state changes
   const updateUrlParams = useCallback(() => {
     const params = new URLSearchParams();
     
     if (debouncedSearchTerm) params.set('search', debouncedSearchTerm);
-    if (debouncedEmailSearchTerm) params.set('emailSearch', debouncedEmailSearchTerm);
+    statusFilters.forEach(status => params.append('status', status));
     if (sortBy) params.set('sortBy', sortBy);
-    if (sortOrder !== 'asc') params.set('sortOrder', sortOrder);
+    if (sortOrder !== 'desc') params.set('sortOrder', sortOrder);
     if (currentPage !== 1) params.set('page', currentPage.toString());
-    if (itemsPerPage !== 10) params.set('limit', itemsPerPage.toString());
+    if (itemsPerPage !== 6) params.set('limit', itemsPerPage.toString());
     
     setSearchParams(params, { replace: true });
-  }, [debouncedSearchTerm, debouncedEmailSearchTerm, sortBy, sortOrder, currentPage, itemsPerPage, setSearchParams]);
+  }, [debouncedSearchTerm, statusFilters, sortBy, sortOrder, currentPage, itemsPerPage, setSearchParams]);
 
   // Fetch data function
   const fetchData = useCallback(async () => {
@@ -134,34 +123,40 @@ export function useDriversData(): UseDriversDataReturn {
     setError(null);
     
     try {
-      const queryParams: DriverQueryParams = {
+      const queryParams: VehicleScheduleQueryParams = {
         search: debouncedSearchTerm || undefined,
-        emailSearch: debouncedEmailSearchTerm || undefined,
+        status: statusFilters.length > 0 ? statusFilters : undefined,
         sortBy: sortBy || undefined,
         sortOrder,
         page: currentPage,
         limit: itemsPerPage,
       };
 
-      const response: PaginatedDriversResponse = await driverService.fetchPaginatedDrivers(queryParams);
+      // Fetch both paginated data and summary concurrently
+      const [response, summaryResponse]: [PaginatedVehicleSchedulesResponse, VehicleScheduleSummary] = await Promise.all([
+        vehicleScheduleService.fetchPaginatedVehicleSchedules(queryParams),
+        vehicleScheduleService.fetchVehicleScheduleSummary()
+      ]);
       
-      setDrivers(response.drivers);
+      setVehicleSchedules(response.vehicleSchedules);
       setTotalCount(response.totalCount);
       setTotalPages(response.totalPages);
       setHasNextPage(response.hasNextPage);
       setHasPreviousPage(response.hasPreviousPage);
+      setVehicleScheduleSummary(summaryResponse);
     } catch (err) {
-      console.error('Error fetching drivers:', err);
-      setError(err instanceof Error ? err.message : 'Failed to fetch drivers');
-      setDrivers([]);
+      console.error('Error fetching vehicle schedules:', err);
+      setError(err instanceof Error ? err.message : 'Failed to fetch vehicle schedules');
+      setVehicleSchedules([]);
       setTotalCount(0);
       setTotalPages(0);
       setHasNextPage(false);
       setHasPreviousPage(false);
+      setVehicleScheduleSummary(null);
     } finally {
       setLoading(false);
     }
-  }, [debouncedSearchTerm, debouncedEmailSearchTerm, sortBy, sortOrder, currentPage, itemsPerPage]);
+  }, [debouncedSearchTerm, statusFilters, sortBy, sortOrder, currentPage, itemsPerPage]);
 
   // Effect to fetch data when dependencies change
   useEffect(() => {
@@ -179,9 +174,19 @@ export function useDriversData(): UseDriversDataReturn {
     setCurrentPageState(1); // Reset to first page when searching
   }, []);
 
-  const setEmailSearchTerm = useCallback((term: string) => {
-    setEmailSearchTermState(term);
-    setCurrentPageState(1); // Reset to first page when searching
+  const toggleStatusFilter = useCallback((status: string) => {
+    setStatusFilters(prev => {
+      const newFilters = prev.includes(status)
+        ? prev.filter(s => s !== status)
+        : [...prev, status];
+      setCurrentPageState(1); // Reset to first page when filtering
+      return newFilters;
+    });
+  }, []);
+
+  const clearStatusFilters = useCallback(() => {
+    setStatusFilters(['active', 'scheduled']); // Reset to default filters
+    setCurrentPageState(1);
   }, []);
 
   const setSorting = useCallback((column: SortColumn) => {
@@ -207,9 +212,9 @@ export function useDriversData(): UseDriversDataReturn {
 
   const clearAllFilters = useCallback(() => {
     setSearchTermState('');
-    setEmailSearchTermState('');
-    setSortByState(null);
-    setSortOrderState('asc');
+    setStatusFilters(['active', 'scheduled']); // Reset to default filters
+    setSortByState('startDate');
+    setSortOrderState('desc');
     setCurrentPageState(1);
   }, []);
 
@@ -219,7 +224,7 @@ export function useDriversData(): UseDriversDataReturn {
 
   return {
     // Data
-    drivers,
+    vehicleSchedules,
     totalCount,
     totalPages,
     currentPage,
@@ -228,7 +233,7 @@ export function useDriversData(): UseDriversDataReturn {
     
     // State
     searchTerm,
-    emailSearchTerm,
+    statusFilters,
     sortBy,
     sortOrder,
     itemsPerPage,
@@ -237,11 +242,15 @@ export function useDriversData(): UseDriversDataReturn {
     
     // Actions
     setSearchTerm,
-    setEmailSearchTerm,
+    toggleStatusFilter,
+    clearStatusFilters,
     setSorting,
     setCurrentPage,
     setItemsPerPage,
     clearAllFilters,
     refreshData,
+
+    // Summary data
+    vehicleScheduleSummary,
   };
 }
